353: Add payment

+++ controllers/admin.js
@@ -168,19 +168,21 @@ exports.postEditProduct = (req, res, next) => {
 
 }
 
-exports.postDeleteProduct = (req, res, next) => {
-    const prodId = req.body.productId;
+exports.deleteProduct = (req, res, next) => {
+    const prodId = req.params.productId;
     Product.findById(prodId).then(product => {
+        if (!product) {
+            return next(new Error("Product not found!"));
+        }
         fileHelper.deleteFile(product.imageUrl);
         return Product.deleteOne({ _id: prodId, userId: req.user._id })
     })
         .then(result => {
             console.log("DELETED !!!")
-            res.redirect("/admin/products");
+            // res.redirect("/admin/products");
+            res.status(200).json({ message: "Success!" });
         })
         .catch(err => {
-            const error = new Error(err);
-            error.httpStatusCode = 500;
-            return next(error);
+            res.status(500).json({ message: "Deleting product failed" })
         })
 }
